% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cm_colors_strong.R
\name{cm.colors.strong}
\alias{cm.colors.strong}
\title{Strong version of cm.colors color palette}
\usage{
cm.colors.strong(n, alpha = 1)
}
\arguments{
\item{n}{Number of color groups}

\item{alpha}{Alpha level}
}
\value{
Character vector of colors
}
\description{
Altered version of cm.colors that uses
full saturation to get stronger colors.
}
\examples{
# Character string output
cm.colors.strong(5)

# Plot to show these
sl <- 21
sx <- seq(0,1,l=sl)
plot(sx,sin(2*pi*sx), cex=5, col=cm.colors.strong(sl), pch=19);points(sx,sin(2*pi*sx), cex=5)
plot(sx,sin(2*pi*sx), cex=5, col=cm.colors(sl),        pch=19);points(sx,sin(2*pi*sx), cex=5)
}
