% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcf.R
\name{gcf}
\alias{gcf}
\title{Make contour plot from data or function using ggplot2}
\usage{
gcf(...)
}
\arguments{
\item{...}{Arguments to be passed to cf_func or cf_data based on 
data type of first argument. If D is given as argument, then it
is passed to cf_highdim.}
}
\value{
Whatever is returned from other function, probably nothing.
Will be a ggplot2 object if using gg=TRUE.
}
\description{
Simpler function for making contours with cf package.
Won't give argument completion, so all must be specified
}
\examples{
gcf(function(x){x[1]^2 - x[2]})
x <- runif(20)
y <- runif(20)
z <- exp(-(x-.5)^2-5*(y-.5)^2)# + rnorm(20,0,.05)
gcf(x,y,z)
gcf(function(x){x[1]^2 - x[2]}, D=3)
}
