% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_4dim.R
\name{cf_4dim}
\alias{cf_4dim}
\title{Plot 2D contour slices of four dimensional functions}
\usage{
cf_4dim(func, over = c(1, 2), nover = 5, nover1 = nover, nover2 = nover,
  low = rep(0, 4), high = rep(1, 4), n = 20, same_scale = TRUE,
  nlevels = 20, color.palette = cm.colors,
  col = color.palette(length(levels) - 1), var_names = paste0("x", 1:4),
  bar = T, key.axes, axes = T, ...)
}
\arguments{
\item{func}{A four-dimensional function to plot contours of}

\item{over}{Indices of the dimensions used for the outer grid}

\item{nover}{Number of grid points for the outer grid dimensions}

\item{nover1}{Number of grid points for the first outer grid dimension}

\item{nover2}{Number of grid points for the second outer grid dimension}

\item{low}{Low input value for each dimension}

\item{high}{High input value for each dimension}

\item{n}{Number of points in grid on each dimension}

\item{same_scale}{Should all contour plots be on the same scale?
Takes longer since it has to precalculate range of outputs.}

\item{nlevels}{Number of levels in contour scale}

\item{color.palette}{Color palette used for contour plots}

\item{col}{The colors for the contour plots}

\item{var_names}{Variable names to add to plot}

\item{bar}{Should a bar be added on right when all on same_scale?}

\item{key.axes}{key for bar plot}

\item{axes}{axes}

\item{...}{Arguments passed to cf_func, and then probably through to cf_grid}
}
\description{
Plots a grid of contour plots.
Each contour plot is a contour over two dimensions with the remaining
two dimensions set to the baseline value.
See cf_highdim for functions with more than 4 dimensions.
}
\examples{
cf_4dim(
  function(x) {x[1] + x[2]^2 + sin(2*pi*x[3])}
)
}
