% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_data.R
\name{cf_data}
\alias{cf_data}
\title{Makes filled contour plot from data without sidebar by interpolating 
with Gaussian process, passes model to cf_func to make contour}
\usage{
cf_data(x, y = NULL, z = NULL, xlim = NULL, ylim = NULL, xylim = NULL,
  fit = "", ...)
}
\arguments{
\item{x}{either just x data, x and y data, or x, y and z data}

\item{y}{either y data, z data, or null}

\item{z}{either z data or null}

\item{xlim}{x limits for the contour plot, will be set to data limits +- 5\% if not specified}

\item{ylim}{y limits for the contour plot, will be set to data limits +- 5\% if not specified}

\item{xylim}{x and y limits for the contour plot}

\item{fit}{Method to fit a model with. Current options are laGP (default)
and mlegp. laGP is faster but might cause trouble.}

\item{...}{passed to cf_func}
}
\description{
Makes filled contour plot from data without sidebar by interpolating 
with Gaussian process, passes model to cf_func to make contour
}
\examples{
x <- runif(20)
y <- runif(20)
z <- exp(-(x-.5)^2-5*(y-.5)^2)
cf_data(x,y,z)
}
\references{
[1] filled.contour R function, copied function but removed part for sidebar

[2] http://stackoverflow.com/questions/16774928/removing-part-of-a-graphic-in-r, answer by P Lapointe
}

