% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_highdim.R
\name{cf_highdim}
\alias{cf_highdim}
\title{Plot contour functions from 2D slices of higher dimensional functions}
\usage{
cf_highdim(func, D, low = rep(0, D), high = rep(1, D), baseline = (low +
  high)/2)
}
\arguments{
\item{func}{Function to plot contours of}

\item{D}{Input dimension of function}

\item{low}{Low input value for each dimension}

\item{high}{High input value for each dimension}

\item{baseline}{Baseline input value for each dimension}
}
\description{
Plots a grid of contour plots.
Each contour plot is a contour over two dimensions with the remaining
dimensions set to the baseline value.
}
\examples{
\dontrun{
# Only use 4 dims of 8 for borehole function
cf_highdim(function(x) TestFunctions::borehole(c(x,.5,.5,.5,.5)), 4)

# Full 8D borehole function
cf_highdim(TestFunctions::borehole, 8)
}

cf_highdim(function(x) {x[1]^2 + exp(x[2])}, D=3)
}
