% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_highdim.R
\name{cf_highdim}
\alias{cf_highdim}
\title{Plot 2D contour slices of higher dimensional functions}
\usage{
cf_highdim(func, D, low = rep(0, D), high = rep(1, D), baseline = (low +
  high)/2, same_scale = TRUE, n = 20, batchmax = 1,
  var_names = paste0("x", 1:D), pts = NULL, average = FALSE,
  average_reps = 10000, ...)
}
\arguments{
\item{func}{Function to plot contours of}

\item{D}{Input dimension of function}

\item{low}{Low input value for each dimension}

\item{high}{High input value for each dimension}

\item{baseline}{Baseline input value for each dimension}

\item{same_scale}{Should all contour plots be on the same scale?}

\item{n}{Number of points in grid on each dimension}

\item{batchmax}{number of datapoints that can be computed at a time}

\item{var_names}{Variable names to add to plot
Takes longer since it has to precalculate range of outputs.}

\item{pts}{Matrix of points to show on plot}

\item{average}{Should the background dimensions be averaged over instead of
set to baseline value? Much slower.}

\item{average_reps}{Number of points to average over when using average}

\item{...}{Arguments passed to cf_func, and then probably through to cf_grid}
}
\description{
Plots a grid of contour plots.
Each contour plot is a contour over two dimensions with the remaining
dimensions set to the baseline value.
Similar to plots created in Hwang et al. (2018).
}
\examples{
\dontrun{
# Only use 4 dims of 8 for borehole function
cf_highdim(function(x) TestFunctions::borehole(c(x,.5,.5,.5,.5)), 4)
# Add points
cf_highdim(function(x) TestFunctions::borehole(c(x,.5,.5,.5,.5)), 4,
           pts=matrix(c(.1,.3,.6,.9),1,4))

# Full 8D borehole function
cf_highdim(TestFunctions::borehole, 8)

# Putting each plot on separate scale
cf_highdim(TestFunctions::borehole, 8, n=10, same_scale = FALSE)
}

cf_highdim(function(x) {x[1]^2 + exp(x[2])}, D=3)

friedman <- function(x) {
  10*sin(pi*x[1]*x[2]) + 20*(x[3]-.5)^2 + 10*x[4] + 5*x[5]
}
cf_highdim(friedman, 5, color.palette=topo.colors)
cf_highdim(friedman, 5, 
           color.palette=function(x) {gray((1:x)/x)},
           nlevels=10)
           
\dontrun{
# Recreate Plate 1 or Figure 1.1 from Engineering Design via Surrogate
# Modelling by Forrester, Sobester, and Keane (2008).
cf_highdim(function(x)TestFunctions::wingweight(x, scale_it=FALSE),
  D=10, low = c(150,220,6,-10,16,.5,.08,2.5,1700,.025),
  high = c(200,300,10,10,45,1,.18,6,2500,.08),
  baseline=c(174,252,7.52,0,34,.672,.12,3.8,2000,.064),
  color.palette=topo.colors, 
  var_names=c('SW', 'Wtw', 'A', 'Lambda', 'q', 'lambda', 'tc', 'Nz', 'Wdg'))
}

# Average over background dimensions, use higher reps to reduce noise.
f1 <- function(x) {x[1] + x[2]^2 + x[3]^3}
cf_highdim(f1, 4, average=FALSE, average_reps=1e2, n=10)
f1b <- function(x) {x[,1] + x[,2]^2 + x[,3]^3}
cf_highdim(f1b, 4, average=FALSE, average_reps=1e2, n=10, batchmax=Inf)

# This was giving bad result
csa()
split.screen(c(2,1))
screen(2)
cf_highdim(f1b, 4, average=F, average_reps=1e2, n=10, batchmax=Inf)
csa()
}
\references{
Hwang, Yongmoon, Sang-Lyul Cha, Sehoon Kim, Seung-Seop Jin,
and Hyung-Jo Jung. "The Multiple-Update-Infill Sampling Method Using
Minimum Energy Design for Sequential Surrogate Modeling."
Applied Sciences 8, no. 4 (2018): 481.
}
