% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_csa.R
\name{csa}
\alias{csa}
\title{Close all open screens}
\usage{
csa(silent = FALSE)
}
\arguments{
\item{silent}{Should the output of `close.screen` not be returned?}
}
\description{
Closes the screens open, which happens
when plotting with `split.screen` is interrupted.
It often happens when there is a error while plotting.
When you try to plot
the next thing it gives an error.
Running this function will reset the plot screen.
It just does `close.screen(all.screens=TRUE)` but is faster to type.
}
\examples{
# Split screen into fourths
split.screen(c(2,2))
hist(rnorm(100))
screen(2)
hist(runif(100))
# Use csa() to go back to normal plotting
csa()
hist(rexp(100))
}
