% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_plot.R
\name{text_plot}
\alias{text_plot}
\title{Make a plot with only text}
\usage{
text_plot(p, x = 0.5, y = 0.5, cex = 2, ...)
}
\arguments{
\item{p}{}

\item{x}{}

\item{y}{}

\item{cex}{}

\item{...}{}
}
\description{
Make a plot with only text
}
\references{
ZNK's answer on https://stackoverflow.com/questions/19918985/r-plot-only-text, retrieved 5/25/2018
}
